Retirada do número de sequencia quando um pacote é recebido. Esta informação já se encontra no nome da mensagem. Bug fix no FileRouter na hora de enviar o sinal de numero de bites quando se dá drop. É necessário fazer um cast antes (feito).

+++ src/FileRouter.cc
@@ -118,7 +118,7 @@ void FileRouter::handleMessage(cMessage *msg)
     } else {
     	EV << this->getName() << " descartando mensagem pois não foi encontrada uma rota para " << pkt->getTo() << "\n";
     	emit(droppedSeqSignalID, pkt->getSeqCount());
-    	emit(droppedLenSignalID, pkt->getByteLength());
+    	emit(droppedLenSignalID, (unsigned long)pkt->getByteLength());
     	delete pkt;
     }
